% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher.R
\name{fisher}
\alias{fisher}
\title{Fisher's method: a technique for meta-analysis.}
\usage{
fisher(datalist, stat = FALSE, p.adj = FALSE)
}
\arguments{
\item{datalist}{input datalist}

\item{stat}{to show stat, by default stat=FALSE}

\item{p.adj}{p values adjustment, by default p.adj = FALSE}
}
\value{
p_value

statistic if stat=TRUE
}
\description{
Fisher's method is used to combine the results from several independence tests
bearing upon the same overall hypothesis (H0). The method sums up the log-transformed p-values
obtained from individual studies.
\deqn{X_{fisher}^2 = \sum\limits_{i=1}^{k}\ln(p_i)}
}
\examples{
dl <- sampleData(p=100)
fisher(dl, stat = FALSE,p.adj = FALSE)
dl <- sampleData(p=100)
fisher(dl, stat = FALSE,p.adj = 'fdr')
}
