% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lancaster.R
\name{lancaster}
\alias{lancaster}
\title{Lancaster's Method: a technique for meta analysis}
\usage{
lancaster(datalist, stat = FALSE, p.adj = FALSE)
}
\arguments{
\item{datalist}{input datalist}

\item{stat}{to show stat, by default stat=FALSE}

\item{p.adj}{p values adjustment, by default p.adj = FALSE}
}
\value{
p_value

statistic if stat=TRUE
}
\description{
Lancaster's method generalized Fisher's method by assignning different
weights using the additivity of chi-squared distribution.
\deqn{T = \sum\limits_{i=1}^{k} [X_{n_i}^{2}]^{-1}(1-p_i)},
where \deqn{[X_{n_i}^{2}]^{-1}} is the inverse cumulative chi-square distributon function with ni degrees of freedom
}
\examples{
dl <- sampleData(p=100)
lancaster(dl, stat = FALSE,p.adj = FALSE)
dl <- sampleData(p=100)
lancaster(dl, stat = FALSE,p.adj = 'fdr')
}
