% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximum_pvalue.R
\name{maxP}
\alias{maxP}
\title{Maximum p-value (maxO): a technique for meta analysis}
\usage{
maxP(datalist, stat = FALSE, p.adj = FALSE)
}
\arguments{
\item{datalist}{input datalist}

\item{stat}{to show stat, by default stat=FALSE}

\item{p.adj}{p values adjustment, by default p.adj = FALSE}
}
\value{
p_value

statistic if stat=TRUE
}
\description{
The maxP method takes maximum p-value as the test statistic.
It follows a beta distribution with degrees of freedom \deqn{\alpha = K} and \deqn{\beta = 1}
under the null hypothesis
}
\examples{
dl <- sampleData(p=100)
maxP(dl, stat = FALSE,p.adj = FALSE)
dl <- sampleData(p=100)
maxP(dl, stat = FALSE,p.adj = 'fdr')
}
