% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_matrix.R
\name{p_vector}
\alias{p_vector}
\title{Computing a P values Vector for a Dataframe}
\usage{
p_vector(df, stat_test = TRUE)
}
\arguments{
\item{df}{Input dataframe}

\item{stat_test}{statistical tests applied to each biomarker in a data frame. Default is false}
}
\value{
A vector of P-values

A vector of statistical tests if stat_test = TRUE
}
\description{
Computing a P values Vector for a Dataframe
}
\examples{
dl <-sampleData()
df <- dl[[1]]
p_vector(df, stat_test=FALSE)
}
