% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_stouffer.R
\name{weighted_stouffer}
\alias{weighted_stouffer}
\title{Weighted Stouffer's Method: a technique for meta-analysis.}
\usage{
weighted_stouffer(datalist, stat = FALSE, p.adj = FALSE)
}
\arguments{
\item{datalist}{input datalist}

\item{stat}{to show stat, by default stat=FALSE}

\item{p.adj}{p values adjustment, by default p.adj = FALSE}
}
\value{
p_value

statistic if stat=TRUE
}
\description{
weight is the square root of the sample size, wi
\deqn{Z_{weighted~Stouffer} = \frac{\sum_{i=1}^{k} w_i Z_i}{\sqrt{\sum_{i=1}^{k} w_i^2}}}
}
\examples{
dl <- sampleData(p=100)
weighted_stouffer(dl, stat = FALSE,p.adj = FALSE)
dl <- sampleData(p=100)
weighted_stouffer(dl, stat = FALSE,p.adj = 'fdr')
}
